Add files via upload
Removes individual threading to enact a turn based scenario. The user is prompted to select the number of troops per side, as well as who will go first. Hit counts are being added, but are currently buggy. The desired fix is to have a larger scoreboard, as well as turn based scoring in addition to an overall hit counter.